[DoNotCarryForward] [ozone/wayland] Fix flickering problems with Weston.
The wl_surface_damage method was used in a wrong manner -
if a damage region is not supplied, the wl_surface_damage
still must be provided with the width and the height of the buffer
to update the surface with a new content.

This happens when partial swap is disabled and there is no damage
region provided.

What is more, instead of using wl_surface_damage, start using
wl_surface_damage_buffer, as long as the damage region corresponds
more to the buffer region contents rather than the surface itself
(and which is claimed to be a better approach by the Wayland
documentation)

Bug: 891188
Change-Id: Iff3a4eb50eae1a113818ef7881b6529b906f4820